--- .changeset/pink-pumpkins-appear.md
@@ -1,7 +1,0 @@
----
-'@cypherock/cysync-core': patch
-'@cypherock/cysync-ui': patch
-'@cypherock/cysync-desktop': patch
----
-
-ui for reminder inheritance component
