+++ packages/ui/src/assets/images/common/index.ts
@@ -13,12 +13,6 @@ import bnbChainIcon from './bnb-chain.svg';
 import cardTapAsideImage from './CardTapAsideImage.png';
 import checkIcon from './check.svg';
 import circledCheckIcon from './circled-check-icon.svg';
-import clockBgDefault from './clock-bg-default.svg';
-import clockBgHover from './clock-bg-hover.svg';
-import clockHover from './clock-hover.svg';
-import ClockInfo from './clock-info.svg';
-import clockDefault from './clock.svg';
-import clockDisabled from './clockDisabled.svg';
 import closeIcon from './close.svg';
 import confirmIcon from './confirm.svg';
 import goldCopyIcon from './copy-gold.svg';
@@ -150,10 +144,4 @@ export {
   UniSwapLogo,
   createWalletGraphics,
   recoverWalletFromSeedphraseGraphics,
-  clockBgDefault,
-  clockBgHover,
-  clockDefault,
-  clockHover,
-  clockDisabled,
-  ClockInfo,
 };
